+++ src/main/java/insane96mcp/survivalreimagined/module/experience/feature/OtherExperience.java
@@ -16,12 +16,6 @@ public class OtherExperience extends Feature {
 	@Config(min = 0, max = 512)
 	@Label(name = "Bottle o' Enchanting Bonus XP", description = "Bottle o' enchanting will drop this more XP. Experience is still affected by 'Global Experience Multiplier'\nCan be set to 0 to make Bottle o' enchanting drop no bonus experience")
 	public static Integer xpBottleBonus = 18;
-	@Config(min = 0)
-	@Label(name = "Anvil Repair Cap", description = "Set the cap for repairing items in the anvil (vanilla is 40)")
-	public static Integer anvilRepairCap = 1024;
-	@Config
-	@Label(name = "Remove rename cost", description = "Removes cost of renaming items in Anvil")
-	public static Boolean freeRenaming = true;
 
 	public OtherExperience(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -36,7 +30,4 @@ public class OtherExperience extends Feature {
 			ExperienceOrb.award((ServerLevel)xpBottle.level, xpBottle.position(), xpBottleBonus);
 	}
 
-	public static boolean isFreeRenaming() {
-		return isEnabled(OtherExperience.class) && freeRenaming;
-	}
 }